+++ domain_alias/src/DomainAliasValidator.php
@@ -40,16 +40,28 @@ class DomainAliasValidator implements DomainAliasValidatorInterface {
       }
     }
     // 3) Check that the alias doesn't contain any invalid characters.
-    $check = preg_match('/^[a-z0-9\.\+\-\*\?:]*$/', $pattern);
-    if ($check == 0) {
-      return $this->t('The pattern contains invalid characters.');
+    // Check for valid characters, unless using non-ASCII domains.
+    $non_ascii = \Drupal::config('domain.settings')->get('allow_non_ascii');
+    if (!$non_ascii) {
+      $check = preg_match('/^[a-z0-9\.\+\-\*\?:]*$/', $pattern);
+      if ($check == 0) {
+        return $this->t('The pattern contains invalid characters.');
+      }
     }
-    // 4) Check that the alias is not a direct match for a registered domain.
+    // 4) The alias cannot begin or end with a period.
+    if (substr($pattern, 0, 1) == '.') {
+      return $this->t('The pattern cannot begin with a dot.');
+    }
+    if (substr($pattern, -1) == '.') {
+      return $this->t('The pattern cannot end with a dot.');
+    }
+
+    // 5) Check that the alias is not a direct match for a registered domain.
     $check = preg_match('/[a-z0-9\.\+\-:]*$/', $pattern);
     if ($check == 1 && \Drupal::service('domain.loader')->loadByHostname($pattern)) {
       return $this->t('The pattern matches an existing domain record.');
     }
-    // 5) Check that the alias is unique across all records.
+    // 6) Check that the alias is unique across all records.
     if ($alias_check = \Drupal::service('domain_alias.loader')->loadByPattern($pattern)) {
       /** @var \Drupal\domain_alias\DomainAliasInterface $alias_check */
       if ($alias_check->id() != $alias->id()) {